Basic tests for ignoring 'secure' cookies
https://tools.ietf.org/html/draft-west-leave-secure-cookies-alone
defines some new behavior for cookies whose 'secure' flag is set. These
will be ignored completely when set from a non-secure page.

Currently, this behavior is only implemented behind a flag in Chrome,
but Firefox and Chrome stable are expected to shift to this behavior in
the relatively near future.